update(compliance-api): AlertController.php

+++ app/Http/Controllers/Api/V1/Compliance/AlertController.php
@@ -99,7 +99,7 @@ class AlertController extends Controller
 
         $results = $this->alertTriageService->bulkResolve(
             $validated['alert_ids'],
-            auth()->id(),
+            (int) auth()->id(),
             $validated['notes'] ?? null
         );
